requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.7.2
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
plotly==5.17.0
gunicorn